Removed the old WinMain.
`WinMain` wasn't very cross platform. It also isn't needed in Windows and actually has some drawbacks. It's not needed because you can call `GetModuleHandle` with a null parameter to get the `HINSTANCE` of the program calling the current function, or more specifically, get the `HINSTANCE` passed through `WinMain`. Its primary drawback is the lack of ability to use a console.