Added some nice utilities for safety.
A basic `Ensure` statement, acts similar to an assert, this is only enabled during debug mode. If the statement fails then it prints a message to `stderr` then triggers a breakpoint.

`NonNull` and `Nullable` are similar to the similarly named annotations in Java, they basically do nothing, mostly for signifying something that is guaranteed to be true.

`NotNull` is a wrapper class that validates taht something is not null. This relies on the `Ensure` statement, so it only does something in debug mode. In release mode it should be completely optimized away, for that matter, it is probably mostly gone in debug mode as well.