Added some basic defines for other keywords
`flags` refers to enum, it is supposed to be analogous to the `[Flags]` annotation (I forget what they call it in C#) in C#.

`interface` is supposed to represent a class that is a pure interface. I am a bit loose about this because of multiple inheritance, specifically I mean that I do also declare some complete methods, mostly things to prevent copying (copy and assignment operators).