Added TUID
Tau Unique ID is similar to a UUID except that it is guaranteed to be unique within a given run of a program. This is done by including an incremental 32 bit number within itself. It may be possible to reverse the number and get back a seed, but that's not really a priority in this. I may modify it a bit later to drop bits to make it impossible to reverse the seed.