Added a whole bunch of GL stuff.
These are mostly OOP wrappers around OpenGL functions. There currently are the GL Program, the shader and the VAO and VBO/IBO. The shader is also an implementation of IShader. This is because I intend on adding DirectX support in the future and wanted to make it as easy to switch between as possible. I currently don't have much experience with DirectX so most of this is based off of the OpenGL API, although from what I have seen of DirectX, I do prefer its shader model more.