Lots of changes to the engine
Removed the binary reader/writer as they didn't serve much use.
Made the vector maths more inline (moved the code to the header).
Added `astdio`, a series of sprintf functions that dynamically allocates strings.
Made `NonNull` `constexpr`, thus encouraging the compiler to throw errors at compile time and leave only raw pointer accesses in the output.
Changed the concept of the engine being a DLL to it being a static library, then the game is a DLL which is loaded into a launcher. This allows for hotswapping of the game code during debugging. Having the engine be static also allows for significant optimizations and unnecessary code removal.
Created a multithreaded rendering pipeline, now it is not an immediate mode rendering, instead it adds instructions to a buffer which another thread will read and execute. I have a code review for it at, but thus far no one has reviewed it (https://codereview.stackexchange.com/questions/219266/multithreaded-opengl-rendering-pipeline).
Added text rendering.