Added basic IFile
Added a basic IFile class. This is used to allow for a universal API for reading and writing to files. This can be overloaded to support any filesystem, including things like compressed archives.

There are also some additional incomplete abstractions for the standard C file API, as well as the Win32 API.