Moved from DLL to Exe
I was never particularly fond of using TauLauncher as a wrapper around a game. As such I've finally moved away from using that. TauEditor is now a full application instead of being a DLL with various exports.

I have not fully ported all of the old code over yet. I'm switching to using an `Application` interface for the runtime. The header includes the `main` function implementation, forcing the client to wrap all of their code in an `Application`. They could technically roll their own implementation, but the `Application` interface is being designed to be as clean and easy to use as possible.

In its current state a window is created and the text handler is up and running, but currently isn't working for some unknown reason.