Fixed bugs, added exceptions
Fixed various bugs. Rearranged code. Added an exception system.

Unlike C++ exceptions (or even worse, Java exceptions), tau's exceptions are only thrown in exceptional cases. When one is thrown it is unrecoverable. There is no stack trace, and no inherent message; just an object containing a unique ID. Using the `ExceptionDispatcher` (basically `EventDispatcher`, but for exceptions) you can invoke a specific handler. From there a message _can_ be formed. 
A major push for this is the fact that various logging strings in code make it *much* easier to decompile. By separating away the throwing of the exception and where the message is created it makes it drastically more difficult to debug. Especially since in the handler it checks for a `RunTimeType` (which is just an extraction of `EventType` for general use), which has a non static ID making it very difficult to correlate to a specific throwing of an exception.