Drastically redid buffers and shaders and uniforms.
There is a now a specific uniform type. The Uniform type is templated and mapped to specific types.

A shader program now contains all shaders. It designed around the shader pipeline of Vertex->TesselationControl->TesselationEvaluation->Geometry->Pixel.

Buffer descriptors are now different from vertex arrays. Every buffer has a descriptor. You can set the layout of buffer by specifying the type, size and offset are automatically computed. When a buffer is loaded into a vertex array it sets up all of the attributes, of course this only occurs with OpenGL, with other API's it's typically a No-Op.