Added uSys type, added array iterators
Added `uSys` type, which is `::std::size_t`, as well as a signed variant (this is determined manually based on the compiler). Added iterator to DynArray. Added some simple error exit macros to Safeties, these are designed around the philosophy of passing an type pointer to receive any errors the function may produce.