Began work on DirectX10
Began implementing and testing DirectX10. I'm setting aside DirectX9 for the time being, it seems a bit dated for the code base, it seems to primarily use the fixed function pipeline.

Began implementation of the rendering context, it can now  clear and swap a frame. Will have to refactor the code so that the swap chain is its own high level concept, this will be especially important for the future as DirectX11, DirectX12, and Vulkan use the same concept.