Reimplemented VertexArray's, added InputLayout
Completely redid VertexArrays, for DirectX at least. Vertex Arrays now rely on an Input Layout which is used by the shaders. The difference between an InputLayout and BufferDescriptor is that an InputLayout doesn't care about how many buffers are used or which offsets they are at. When loading the shader into the device, the only thing that matters is the format of the data. When loading the actual buffers it starts to care about what the offsets and strides of each buffer are.

Also did some minor refactoring and did some work to improve RunTimeType.