Finished DirectX10 shaders and buffers, removed input layouts
Shaders and buffers for DirectX10 are now finished and functional. Input Layouts were removed, this was done because I came to the conclusion - even though most people object - that DirectX Input Layouts and OpenGL Vertex Arrays are the exact same thing. I will probably add a wiki page talking about, mostly so that I don't forget. Basically what caused me to redo it was I realized that `inputSlot` actually referred to which buffer, as such Input Layouts cannot generically describe a shaders input with a variety of input formats (buffer layout patterns), as such an Input Layout is now bound to the Vertex Array. In the process of doing this I also removed multiple contexts, they never really worked for what I intended and weren't really worth the extra effort.

Vertex Arrays now - once again - store a reference to all of its containing buffers. This had to be done due to some crashing and what amounted to a use after free, that manifested itself in the driver and not my application, so that was a fun several hours. I've added a comment so that I don't ever try to change it again.

Still have to update some of the code everywhere in TauEditor, the vertex array builder needs a shader now. OpenGL doesn't need it so it will still run, but something that needs to be done.