Implemented and tested `FixedBlockArenaAllocator`
The `FixedBlockArenaAllocator` lets you allocate a large number of objects very quickly, and then release them all at once. Because this is C++ you still have to manually deallocate each one so that the destructor is invoked, but the deallocation doesn't actually free the data. Once you are finished with your workload you can simple release the allocator and all the memory will be freed back to the system.

This allocator relies on the lowest level allocator available in Win32, `VirtualAlloc`. This gives the most performance and some pretty important benefits. With a normal expanding allocation strategy seen in high level code, like a vector, you allocate your memory, and when you need more you allocate a larger block and move over the data. With `VirtualAlloc` we can instead reserve a near infinite amount of memory, and only commit the pages when they're needed. Doing this has some nice benefits; the primary one being you don't have to move over the old data once you run out of space. Because the reserved pages are contiguous, where one ends another will begin, even if they aren't committed together. Getting into why is needlessly complicated, but basically comes down to kernel, GDT, protected mode, and paging; if you're curious you can learn about all of those on OSDev.

There are also some special tracking options. Currently there are just 2 (well, 3, one of which is no tracking) options, allocation counting and double delete tracking. The difference being that `AllocationTracking::Count` simply keeps track of the number of allocations compared to the number of deallocations. This is all done with a single number. If the value is negative you deallocated an object too many times. If the value is positive you forgot to deallocate some objects. `AllocationTracking::DoubleDeleteCount` keeps track of the number of allocations compared to the number of deallocations, but also stores information with each allocation specifying the number of times it was deallocated. This is just a `size_t` stored directly before the allocation, so it's not too hard to view it, that being said, there is no direct way to access that information with the allocators. The allocator will then keep track of the number of times an object was deleted twice and the number of times and object was deleted more than once. The difference between these two values is that the former only tracks double deletes, it doesn't keep track of any deletions after that; the latter on the other hand does keep track of every time an object was deallocated more than once, this does include double deletes.

To the best of my knowledge the allocator has been fully tested. The memory is valid and can be accessed. When it runs out of memory it simply returns null. The tracking allocators do seem to track everything correctly.