Converted all GL buffer builders.
All OpenGL buffer builders follow the new pattern. The pattern has also slightly changed; now you do once again get an IBufferBuilder, but it is now a single reference allocated once. The builder then lets you create the buffer from the args, the difference is that now it can create a raw pointer, a raw pointer with a `TauAllocator`, a `shared_ptr`, a `NullableReferenceCountingPointer`, or a `NullableStrongReferenceCountingPointer`. This was done as the tau reference counting pointers finally felt like they were in a usable state, also given that this would happen at some point in the future no matter what, I decided that today was as good as any day, at the very least it means less work down the road.