Added Uniforms for DirectX10
Added uniforms for DirectX10. Need to do some work on automatically transposing matrices depending on the API. For some reason a GLM orthographic view matrix only works with OpenGL, but an XMMatrix orthographic view matrix works with both OpenGL and DirectX10.

Added some stuff to ResourceLib to allow for selecting a resource based the rendering mode. It's far more generic than that, but that's the original purpose I built it for.