Added Shader Bundle Parsing
Shader bundles are going to be used to map uniforms, inputs, and outputs from a shader to a program. In DirectX every shader stage has its own indices for uniforms (constant buffers), whereas in OpenGL you can't use the same indice twice in a shader program. Along with that OpenGL only allows for explicitly setting uniform buffer binding, and individual buffer locations in GLSL 420 and 430 respectively (IIRC). In older versions you would have to manually query their index from the OpenGL driver. A similar setup is required for inputs and outputs.

To ensure the maximum ease between DirectX and OpenGL I've opted for shader bundle information files. Instead of setting in code the various names of uniforms and inputs/outputs for each shader they can now be set in this shader bundle information file. Then the respective tau driver can handle its information how it wants to leave mostly everything opaque to the end user. Another benefit of this is it will make it easier to load shaders. Where you currently have to create a shader program, and then each of your shader stages, then add them to the shader program, then link and compile everything, instead in the future it should be as simple as specifying the path the shader bundle file and the parser will handle the rest. Of course this does now require creating the shader bundle files, but the amount of work for that is rather trivial compared to the actual amount of code that used to be required. Along with that you will be able to swap out entire shader programs without having to rebuild the entire program, while there were definitely methods of doing this in the past, this will make it a bit cleaner.

Some things that still need to be done include adding input and output to the lexer and parser, this is mostly about adding a string check and a value to the token enumeration, then specifying it as a valid block type in the parser. The actual tau graphics drivers still also need to be updated to handle this. That should mostly involve creating a visitor. Right now there is one example visitor that can be used to print out the entire file. If everything worked correctly the output file should be no different than the input. There may be some minor differences, namely the indent (which can be configured currently, but only in constant spaces), and the name of certain blocks, while the files do support both HLSL and GLSL naming standards (as you will be using them with specific API's and may want to follow their terminology) the printer just generalizes to the ones I prefer (HLSL Pixel and GLSL Tessellation) as I feel these are the most descriptive (technically Fragment is more accurate, but pixel more clearly explains what is happening).