Working on getting DX10 to work with TauEditor
It's still not rendering, but due to some instability on my computer wanted to push either way. There are several other major changes that have occurred. 

The orthographic viewing matrix is now built from the DirectX style, thus (0, 0) is in the center of the screen. This feels better, but does require some minor changes, particularly around UI. The Console layer has been updated to once again properly render text.

Finally fixed the issue with `tauExit` not properly storing the exit code. Turns out that in the destructor of the application the window was destroyed which invoked `DestroyWindow` which triggered the `WindowProc` with `WM_DESTROY`, which posted `0` for an exit code. This has been changed such that Window will no longer set an exit code on close. The assumption now is that the default exit code is already 0, if a specific exit code is needed one should call `tauExit` manually. The close function still calls `tauExit` in case it hasn't been called yet (ie, the close button was pressed) so that `tauShouldExit` returns true.

Added some more specializations for GLM. Mostly some stuff for transposition, and some projection specializations.

Minor updates to the shader bundle parser, it still doesn't do anything formal yet though.

Added shader stages into uploading of textures. This is required for DirectX. Converted the texture builders to the new format. Removed individual depth and stencil options from framebuffers, all graphics implementations seem to use the compounded Depth24Stencil8 format.

Redid some things in how window handles are looked up. Basically switched it from a linear lookup array to an AVL tree. This wasn't really necessary, but it seemed valuable if you have lots of windows for some reason. Also I was bored and tired of debugging DirectX.