Abstracted Depth-Stencil States
Skybox had some old OpenGL code in it still, to resolve this I had to redo the entire concept of depth testing state in the rendering context. The old system hasn't been fully removed yet. The new system is much more like DirectX10's way of doing things. Now you set the current depth stencil state in the context, the context then applies it in its own way; for DirectX this is done with OMSetDepthStencilState, whereas OpenGL calls all of the individual state functions to fully set everything up. Now if you want to tweak one state setting you have to copy the default states parameters and build a new state, DirectX10 allows for up to 4096 _distinct_ states (identical states don't affect this count, they just return the same instance), that being said creating two identical states created within Tau Engine are going to have a unique instance.

Renamed `Ref` to `CPPRef`, made `ReferenceCountingPointer` the de facto `Ref`. 

Shader bundle parsing is somewhat broken. There is a check to ensure that a unforms block cant be outside of a stage block, but if a different block (not that there are currently any valid blocks that would fit this) were to precede the uniforms block, then the `_currentBlock` state would show that preceding block instead of the parent block. The solution to this is to distinctly identify the current block, and the parent block; maybe not even the current block.