Lots of changes
Began reworking the API to more closely match DirectX11. This is done by moving the creation of objects out of the rendering context and into the graphics interface. This more closely resembles DirectX11's approach where the device creates the objects, but you set the actual state within the device context.

Began adding DirectX11 support, ironically its abstractions do not yet support the newly designed interface. I was mostly concerned with getting it working with testable API's.

Added small string optimization.
Cleaned up TauEditor.
Removed support for OpenGL 2.0.

Removed `RenderingPipeline`, it was intended to simulate command queues, but ultimately became something else, that was somewhat pointless. In the future I will be creating a Command Queue abstraction which will require some specialization for OpenGL. The basic principle will be that in Vulkan (and DX12 IIRC) it will just be a wrapper around the existing type, but for OpenGL and DirectX it will queue up a list of commands, this will likely be a tagged union which points to an abstract graphics object (like a buffer or texture) and then some tag to signify which action to perform. There hopefully will be some multithreading potential for this, even in OpenGL, albeit not on the actual rendering calls.

Added lots of transposed matrices to use with DirectX. I know that the memory layout should be the same between types, but for some reason DirectX only works when I transpose the matrix, so now I cache the transposed matrix basically everywhere. Maybe in the future I'll figure out how to get everything working with a single matrix. There seems to be some stuff relating to pre and post multiplication, and it does appear that hlsl supports both. I don't know if opengl supports both, since I have tried reversing the order of operations in the past to no effect.

Countless other changes.