Fixed matrices and skybox
Switched to pre-multiplication for matrices, or maybe it's post-multiplication. Fixed the skybox, there were 2 issues, the first was that the vertices needed to be set in the initial buffer for the skybox. The second being cubemaps were setup incorrectly for DirectX, mostly had to change the indexing pattern for the subresources, previously it was mapping the index in the initial buffer array when it needed to map the index in the descriptor array.