Began VR, fixed lighting in DX10, add buffer safeties
Began implementation of VR. If the config as the vr key set (which it is by default) it will attempt to launch in VR mode. If the OpenVR fails to detect the SteamVR runtime or if an HMD is not connected it will fallback to the usual play mode. Right now VR is hardly functional. At best it might work in OpenGL, for some reason it's triggering an OpenGL error when sending the framebuffer over. Turning of the camera has been tuned, but still has significant cumulative error issues, for some reason the position isn't quite ready yet, at least that's my best guess as to why the model isn't showing up.

Fixed lighting for DirectX. The issue was with modification of the constant buffer for the specular exponent. The engine now properly calls the `beginModification` method to map the buffer. Because of this I chose to add many safeties to the buffer so that this shouldn't happen again. The only thing further to implement that I can think of would be destruction checks to ensure that the buffer isn't bound or in the process of being modified.

I also added a code breakpoint in the debug implementation of the `tauThrowException` function. This was done because the stacktrace doesn't show up with the current exception model, this is due to a flaw where exceptions aren't handled until the end of an update cycle. Now exception handling is still the same but if you're running in debug mode (or more correctly, not in production) a breakpoint should be triggered when the exception is thrown, from there it is possible to follow the stacktrace to fully understand the source of the cause.