Fixed crashing and text rendering.
The crashing was ultimately due to the fact that Null textures were set to not have a VTable despite being a concrete class. Text rendering wasn't working because the rendered framebuffer was at the near frustum, moving it to the back temporarily fixed the issue. Should probably implement some way of copying over the framebuffer's depth buffer.