Various minor fixes for a release.
The engine now looks for a local directory for all of its resources. Fixed some various bugs for when it fails to find a resource.

Added credits to the readme. This appears to only be required by FreeType, but it is probably a good idea to list all of the libraries I use.