Reworking rendering and shaders.
Began reworking shader loading. The new model has a visitor which extracts all the useful information, then a per API program should take that data and actually load the shader. This is different from the last version in which each API was responsible for writing their visitor to extract the information, but because the information is pretty much the same for all of the API's it was easier and cleaner to have a single visitor which is used purely for extracting the important data from the AST into some simpler, less branchy, structs.

Before implementing the IShaderProgramBuilder I decided to look into how exactly I would want these to be used. I've decided that it would be nice to be able to swap out the entire shader program, or individual shader stages from within the renderer. I can either ditch the concept of a shader program outright, thus relying on either OpenGL 4.1 or an ARB. Both of these are a decade old, thus the oldest Nvidia GPU's with support either the GeForce 300 series (which only supported the complete core of OpenGL 3.3, but also supports Direct3D 10.1, which has was originally designed with separate shaders in mind) or the GeForce 400 series (which had full support for OpenGL 4.6). The oldest GPU's I hear people actively using these days are the GeForce 700 series. According to the steam hardware survey only 2.07% of users use a 750 Ti, with anything older having less than 1% market saturation. Also according to the survey 92.29% of users have DirectX12 compatible GPU's, with a drastic falloff the older the API, at DirectX9 the market saturation is 0.07%. As such I am dropping support for DirectX9, and any OpenGL version less than OpenGL 4.2.

I also redid the parser for shader bundles. Now the file has blocks for each of its supported API's. The textures block is no longer within the Uniforms block. A Common Render Model has been implemented to make it easier to develop a shader that can be used as an adhoc replacement. You can now specify these CRM types instead of raw integer indices, they just immediately converted to their integer equivalent. If I add a binary format I'll probably want to store the CRM values within the AST to alleviate any issues in CRM changes in the future.

Made many improvements to the various strings. Can't remember why, but they're their now. 
Implemented FixedBlockAllocator.
Began a rough implementation of a renderer.
Added some additional utilities to the VFS.
Improved move semantics for various TauUtils objects.

Doesn't compile right now, just wanted to commit it to the experimental branch.