Lots of refractoring
Removed the concept of a general buffer in favor a vertex buffer. Removed the individual builder types for buffers in favor of a single builder with specialized functions for each buffer type. Removed the redundant bind functions for buffers, they weren't ever used. Uniform buffer binding is still available. The error checking for binding has also been removed.

Added some utilities to the VFS for extracting information from a path.

Changed ArrayList from accepting a number of pages to reserve to accepting a max number of elements.

Reworked shaders partially. This was initially done in favor of separate shader programs similar to DirectX. That was until I realized that OpenGL went from having individual shaders that are attached to a shader program, to having individual shader programs that are attached to a pipeline object. Thus IShaderProgram will still continue to exist in the code.

Changed up vertex arrays again. Now there is an input layout which stores information about how a set of buffers should be decoded (it also stores the offsets and strides as these are expected to be identical for all buffers, may have to change this a little in the future). Vertex arrays now a simply an array of buffers, with the backend implementations storing the raw buffers of the underlying API types. Legacy versions of OpenGL are still somewhat supported; the input layout is mostly just an empty wrapper.