Added Vector2f
This requires a slightly different LLVM IR interface, this is because I was passing a struct into the function, for some reason this doesn't have proper code gen, thus I've changed it to more accurately represent the code it is generating. It also needed a storage pointer because C was attempting to receive the function output from the A register, while the function was attempting to return the value in the XMM0 register.