Began reworking buffer resources.
This is part of the ongoing conversion to an API that more closely matches that of DirectX12. Currently working on changing buffers to be the more low level construct of a resource, with views that describe how to interpret the data.

Part of this work is required to get TextureView's to work correctly as textures will also need to be dissolved into raw resources.