Began adding descriptor heaps.
The entire purpose of this transition is to make the API more analogous to DirectX12 or Vulkan. One of the required changes is how resources are bound to the shader pipeline. Instead of simply passing in some ID3D12ShaderResourceView's, you have to allocate them in a block, and the use that block as the upload. 

To facilitate this feature I've added DescriptorHeap's and DescriptorTable's. In DirectX12 descriptor tables are simply a index into the heap, but due to the backwards compatibility to older API's I've decided to make it an opaque void*. In OpenGL this is expanded into an entire class storing an array of resources. For performance reasons this array is allocated as a large block and the individual resource views are placement new'd into that block. This quite drastically changes the builder for ITextureView and IUniformBufferView. Now they simply return a raw pointer that does not have to be destroyed. As best as I can tell, the main thing you need to care about is that descriptor table, thus the individual views are relatively useless, and don't need reference counting.

Currently only OpenGL is implemented.