Changed alignment settings for OpenGL resources.
Previously the map count of `GLResourceBuffer` and `GLResourceTexture` was aligned to the 64 byte boundary. This was to ensure atomicity. This left the problem that it was a pointer to volatile data, thus to delete it with the proper expression (ie the delete operator accepting a custom alignment) I would have to remove the `volatile` keyword using `const_cast`. This was technically undefined behavior. On the other hand a raw delete expression is able to ignore the volatile nature of the data. The data is required to fit within a cache line to be used atomically, but it does not have to be aligned to the 64 byte boundary. Instead it need only be aligned to the 8 or 4 byte boundary (depending on the architecture word size of size_t) to ensure it remains within a cache line. This would still require some alignment from new. Thankfully according  ISO/IEC 14882:Working Draft N4861 §6.7.5.4.1.3.3, the allocation version of the new operators will return a pointer to memory that is aligned to the size of the type (as long as the types size doesn't exceed the new extended alignment as defined by §6.7.6, which size_t _should_ be within). Thus I'm removing the explicit alignment, allowing me to use the delete expression, allowing the variable to be volatile without having a second non-volatile version.