Some minor changes to TauUtils and allocators.
Changed a C style cast to a `reinterpret_cast`.

Changed the fallback definition of `iSys` to `::std::ptrdiff_t`. As far as I can tell `ptrdiff_t` is supposed to be the signed counterpart to `size_t`. There do seem to be some minor differences, but this should reasonably work given that a `size_t` is supposed to be able to handle a regular pointer (non-member), and `ptrdiff_t` is supposed to be able to store the difference between two pointers, thus they should both be the architecture word size.

Added the ability to commit multiple pages with `PageAllocator`.

Added some utility functions for creating alignment to `_TauAllocatorUtils`. These are specifically for determining an alignment value, and not for actual alignment. i.e. `_alignTo` will round up to a specified alignment. `_nextPowerOf2` can be used to ensure alignments are strictly a power of 2.

The fixed block allocators now take in `allocPages`, a number used to specify how many pages should be committed at a time. This is done to reduce the number of calls to the kernel. By default it will now commit 4 pages at a time, which will increase the initial memory usage by a factor of 4, but should also reduce the number of kernel calls by an equivalent factor of 4.