Continued work on Command Lists
Began adding command queues, in OpenGL this just translates a command list into raw OpenGL calls.

Added a state manager for OpenGL. This is used to limit the number of driver calls that have to be made due to redundant data.

Changed OpenGL's texture view implementation. Previously it was just a pointer to the texture resource, now it is a structure also containing the texture target type. Not sure why I didn't do this before, seems obvious in hindsight.