Began working on pipeline states.
In DirectX12 all of the various settings for the pipeline are stored in one homogeneous state. Unfortunately older versions do not operate the same way. Thus I've chosen to implement the pipeline state as a compound of other state objects. This allows for DirectX10 and DirectX11 to set their individual state objects while also allowing DirectX12 to hide away all of the data into a single state object.

Part of this change thus required implementing setting of the pipeline state within the command list. In OpenGL I've chosen to simply set all of the pipeline state attributes once it encounters that command. This takes advantage of the GLStateHelper to ensure that as few actual state changes occur as possible. This is only just beginning to be implemented, and thus I've only completed the blending states.