General conversions, worked on shaders.
Lots of general changes relating to the Objects macros.

Worked on shaders some more. I've decided to go with an opaque model for shader programs, and an invisible model for shaders. In a practical sense you don't really need the individual shaders once they are attached to a program, thus I've removed the high level concept. To help with reducing memory I've added a shader manager to OpenGL to ensure that two identical shaders have the same object. This is more or less a basic hashmap of the path to the shader and the shaders handle.