Reworked reference counting pointers, continued working on command lists
Decided to do some work on reference counting pointers. They now all inherit from a templated base class, this class lacks much purpose, but would allow an application to receive an arbitrary ref pointer and be able to use independent of whether it is a normal ref pointer or a strong/weak pointer. This templated class inherits from a non-templated base class, this is so they have a virtually invokable destructor, this will come up later.

After some conversations I've decided to force reference counting types throughout the engine. This is done as a safety measure, and should have a minimal effect on performance.

To meet this end I needed to do some changes to command lists. Because I store all commands using the underlying data, and because in DirectX12 and Vulkan I won't even be using mocked commands, I needed a way to store the reference count of the object while it is stored inside a command list. To do this I decided to create a free list. 

The principal behind this is that you have an array of ref pointers that need to be freed after the command list has been submitted. The problem is that there are multiple different pointer types being used (specifically the data being pointed to is different), so I needed a way to store them all in the same array and still be able to invoke their individual destructors (which in turn may invoke the underlying data destructors). The obvious solution is a base class for the pointers, but the next problem is that the array will be a fixed size and the interfaces won't fit correctly. The usual solution is to then store an array of pointers and dynamically allocate each pointer and store its interface pointer. This is slow and unnecessary. Instead I've decided not to use an array, per se, but use a custom allocator. This will treat ref pointers as merely untyped blocks of memory. This allocator is a modified fixed block arena allocator, this means each block is the same size (the actual ref pointers should be the same size given they both just store a single pointer to their respective data objects, on most systems this should mean they're the same size) and that they are all released at once, instead of being able to release individual blocks and reallocate in their place. The difference with this one is that you can invoke a reset (which was retroactively added to the FixedBlockAllocators). When the allocator is reset or destructed all of the allocated pointers are interpreted as the base pointer type (iterating along the stream of blocks and then casting each one to the base pointer to ensure a correct stride) and then their destructor is virtually invoked.

Also made various additions to other parts of the command lists.