Continued work on command queues and descriptor tables
Continued working on implementing command queues for DirectX12. Updated OpenGL's command queues and lists to support the `startInstance` parameter.

Did a lot of work on descriptor heaps and tables. I've basically gone with completely emulating DirectX12 and ditched the table objects. Now there are individual heaps for each descriptor type (this is to make supporting Vulkan easier), and instead of being able to create a table you can get a pointer to the head of the heap and the offset stride. This 'handle' (which is just a wrapper nearly identical to `CD3DX12_CPU_DESCRIPTOR_HANDLE`/`CD3DX12_GPU_DESCRIPTOR_HANDLE`, which wraps a number which itself is just an integer representation of a pointer) is then passed to the builder of the various views.

OpenGL's heap currently uses `malloc` instead of `new[]`. This is because I'm allocating a placement array, and I don't want to initialize all the objects at instantiation, thus I need a large buffer of `n` bytes. Using `malloc` felt less clunky than `new u8[]`, which relies on `u8` actually translating to a byte. While this shouldn't be a problem, it is C++, so you never know what disastrous design decision ended up in the compiler or standard library. If I truly wanted to be purist (I don't) I could use `operator new` instead of `malloc` for much the same effect, albeit in a much more clunky form.