Made various improvements to ResourceLib
Added some new or improved functionality to IFile, mostly fully supporting wide and narrow chars, as well as base path + sub path shortcuts.

Also began working on a texture format and some new file reader/writers. The purpose of the reader/writers is to segregate that functionality away from IFile, and to make it easier to optionally support buffering and sparse files.

TauTexture is far from complete. Currently the philosophy is that it will store a texture similar to how it is stored in the GPU. This means storing a width, height, depth/array count, and mip levels. It also has a limited selection of formats it can use. Support for S3TC is not currently planned. Every texture subresource is placed in its own block, potentially with padding between every subresource for improved unbuffered IO. The size effects of this will be mitigated by using sparse files. The block alignment is variable, but will likely end up on the 4K boundary. Most drives use a 256 byte block, so assuming that the file system uses this block size for the minimum valid sparse region, then many portions of the file should end up as sparse data.

A list of all sub resources is at the beginning of the file, this list contains the offset, the compressed length, and the uncompressed length. This list is contiguous and is indexed by the subresource index. The actual textures do not have to be in order by their subresource index.